+++ frontend/src/components/Content.jsx
