+++ frontend/src/components/TableYarina.jsx
