+++ frontend/src/screens/AllProjectsScreen.jsx
