+++ frontend/src/screens/ProjectSpecsScreen.jsx
